task_management/requirement.txt
bcrypt
boto3
botocore
channels
coreapi
cryptography
Django
django-cleanup
django-cors-headers
django-environ
django-extensions
django-import-export
django-jsonfield
django-log-request-id
django-webpack-loader
django-wkhtmltopdf
djangorestframework
docutils
drf-yasg
fpdf
html5lib
hyperlink
img2pdf
mysqlclient
num2words
openpyxl
phonenumbers
Pillow
PyJWT
pyOpenSSL
python-dateutil
reportlab
reportlab_qrcode
requests
s3transfer
simplejson
sqlparse
urllib3
xhtml2pdf
django-simple-history
python-barcode
cmake
face-recognition
dlib
opencv-python
django-debug-toolbar
django-picklefield
django-silk
django-crontab
pandas
matplotlib

django-secure
django-axes